backend/requirements.txt
# FastAPI core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# Google OAuth
google-auth==2.23.4
google-auth-httplib2==0.1.1

# Redis
redis==5.0.1

# Other tools
httpx==0.25.2

